@@ -23,23 +23,33 @@
 
 package org.shadowice.flocke.andotp.Activities;
 
+import android.Manifest;
 import android.animation.ObjectAnimator;
 import android.app.AlertDialog;
 import android.app.KeyguardManager;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.content.pm.PackageManager;
 import android.content.res.Configuration;
 import android.os.Bundle;
+import android.os.CountDownTimer;
 import android.os.Handler;
 import android.preference.PreferenceManager;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.SearchView;
-import android.support.v7.widget.Toolbar;
-import android.support.v7.widget.helper.ItemTouchHelper;
+import androidx.annotation.NonNull;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.lifecycle.DefaultLifecycleObserver;
+import androidx.lifecycle.LifecycleOwner;
+import androidx.lifecycle.ProcessLifecycleOwner;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.appcompat.widget.SearchView;
+import androidx.appcompat.widget.Toolbar;
+import androidx.recyclerview.widget.ItemTouchHelper;
+import android.text.TextUtils;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
@@ -62,10 +72,11 @@ import org.shadowice.flocke.andotp.Utilities.Constants;
 import org.shadowice.flocke.andotp.Utilities.EncryptionHelper;
 import org.shadowice.flocke.andotp.Utilities.KeyStoreHelper;
 import org.shadowice.flocke.andotp.Utilities.NotificationHelper;
+import org.shadowice.flocke.andotp.Utilities.ScanQRCodeFromFile;
 import org.shadowice.flocke.andotp.Utilities.TokenCalculator;
 import org.shadowice.flocke.andotp.View.EntriesCardAdapter;
 import org.shadowice.flocke.andotp.View.ItemTouchHelper.SimpleItemTouchHelperCallback;
-import org.shadowice.flocke.andotp.View.ManualEntryDialog;
+import org.shadowice.flocke.andotp.Dialogs.ManualEntryDialog;
 import org.shadowice.flocke.andotp.View.TagsAdapter;
 
 import java.util.ArrayList;
@@ -82,6 +93,7 @@ public class MainActivity extends BaseActivity
     public static long animatorDuration = 1000;
 
     private static final String INTENT_SCAN_QR = "org.shadowice.flocke.andotp.intent.SCAN_QR";
+    private static final String INTENT_IMPORT_QR = "org.shadowice.flocke.andotp.intent.IMPORT_QR";
     private static final String INTENT_ENTER_DETAILS = "org.shadowice.flocke.andotp.intent.ENTER_DETAILS";
 
     private EntriesCardAdapter adapter;
@@ -98,12 +110,16 @@ public class MainActivity extends BaseActivity
     private ListView tagsDrawerListView;
     private TagsAdapter tagsDrawerAdapter;
     private ActionBarDrawerToggle tagsToggle;
+    private String filterString;
+
+    private CountDownTimer countDownTimer;
 
     // QR code scanning
     private void scanQRCode(){
         new IntentIntegrator(MainActivity.this)
                 .setOrientationLocked(false)
                 .setBeepEnabled(false)
+                .setCaptureActivity(SecureCaptureActivity.class)
                 .initiateScan();
     }
 
@@ -193,6 +209,8 @@ public class MainActivity extends BaseActivity
             }
         });
 
+        ProcessLifecycleOwner.get().getLifecycle().addObserver(new ProcessLifecycleObserver());
+
         if (! settings.getFirstTimeWarningShown()) {
            showFirstTimeWarning();
         }
@@ -209,6 +227,9 @@ public class MainActivity extends BaseActivity
                         return false;
                     case R.id.fabEnterDetails:
                         ManualEntryDialog.show(MainActivity.this, settings, adapter);
+                        return false;
+                    case R.id.fabScanQRFromImage:
+                        openFileWithPermissions(Constants.INTENT_MAIN_QR_OPEN_IMAGE);
                         return false;
                     default:
                         return false;
@@ -275,12 +296,34 @@ public class MainActivity extends BaseActivity
 
         setupDrawer();
 
+        if (savedInstanceState != null){
+            setFilterString(savedInstanceState.getString("filterString", ""));
+        }
+    }
+
+    private void checkIntent() {
         Intent callingIntent = getIntent();
         if (callingIntent != null && callingIntent.getAction() != null) {
-            if (callingIntent.getAction().equals(INTENT_SCAN_QR)) {
+            // Cache and reset the action to prevent the same intent from being evaluated multiple times
+            String intentAction = callingIntent.getAction();
+            callingIntent.setAction(null);
+
+            if (intentAction.equals(INTENT_SCAN_QR)) {
                 scanQRCode();
-            } else if (callingIntent.getAction().equals(INTENT_ENTER_DETAILS)) {
+            } else if (intentAction.equals(INTENT_IMPORT_QR)) {
+                openFileWithPermissions(Constants.INTENT_MAIN_QR_OPEN_IMAGE);
+            } else if (intentAction.equals(INTENT_ENTER_DETAILS)) {
                 ManualEntryDialog.show(MainActivity.this, settings, adapter);
+            } else if (intentAction.equals(Intent.ACTION_VIEW)) {
+                try {
+                    Entry entry = new Entry(callingIntent.getDataString());
+                    entry.updateOTP();
+                    entry.setLastUsed(System.currentTimeMillis());
+                    adapter.addEntry(entry);
+                    Toast.makeText(this, R.string.toast_intent_creation_succeeded, Toast.LENGTH_LONG).show();
+                } catch (Exception e) {
+                    Toast.makeText(this, R.string.toast_intent_creation_failed, Toast.LENGTH_LONG).show();
+                }
             }
         }
     }
@@ -316,16 +359,45 @@ public class MainActivity extends BaseActivity
                 } else {
                     populateAdapter();
                 }
+                checkIntent();
             }
+
+            if (setCountDownTimerNow())
+                countDownTimer.start();
         }
+
+        if (filterString != null) {
+            // ensure the current filter string is applied after a resume
+            setFilterString(this.filterString);
+        }
+
+        View cardList = findViewById(R.id.cardList);
+        if(cardList.getVisibility() == View.INVISIBLE)
+            cardList.setVisibility(View.VISIBLE);
 
         startUpdater();
     }
 
     @Override
     public void onPause() {
+        if(settings.getAuthMethod() == AuthMethod.DEVICE)
+            runOnUiThread(new Runnable() {
+                @Override
+                public void run() {
+                    findViewById(R.id.cardList).setVisibility(View.INVISIBLE);
+                }
+            });
         super.onPause();
         stopUpdater();
+
+        if (countDownTimer != null)
+            countDownTimer.cancel();
+    }
+
+    @Override
+    protected void onSaveInstanceState(Bundle outState) {
+        super.onSaveInstanceState(outState);
+        outState.putString("filterString", filterString);
     }
 
     public void onSharedPreferenceChanged(SharedPreferences prefs, String key) {
@@ -334,11 +406,15 @@ public class MainActivity extends BaseActivity
                 key.equals(getString(R.string.settings_key_split_group_size)) ||
                 key.equals(getString(R.string.settings_key_thumbnail_size))) {
             adapter.notifyDataSetChanged();
+        } else if (key.equals(getString(R.string.settings_key_search_includes))) {
+            adapter.clearFilter();
         } else if (key.equals(getString(R.string.settings_key_tap_to_reveal)) ||
                 key.equals(getString(R.string.settings_key_theme)) ||
                 key.equals(getString(R.string.settings_key_locale)) ||
                 key.equals(getString(R.string.settings_key_enable_screenshot)) ||
-                key.equals(getString(R.string.settings_key_tag_functionality)) ) {
+                key.equals(getString(R.string.settings_key_tag_functionality)) ||
+                key.equals(getString(R.string.settings_key_label_highlight_token)) ||
+                key.equals(getString(R.string.settings_key_card_layout))) {
             recreate();
         }
     }
@@ -357,16 +433,7 @@ public class MainActivity extends BaseActivity
         IntentResult result = IntentIntegrator.parseActivityResult(requestCode, resultCode, intent);
         if(result != null) {
             if(result.getContents() != null) {
-                try {
-                    Entry e = new Entry(result.getContents());
-                    e.updateOTP();
-                    e.setLastUsed(System.currentTimeMillis());
-                    adapter.addEntry(e);
-                    adapter.saveEntries();
-                    refreshTags();
-                } catch (Exception e) {
-                    Toast.makeText(this, R.string.toast_invalid_qr_code, Toast.LENGTH_LONG).show();
-                }
+                addQRCode(result.getContents());
             }
         } else if (requestCode == Constants.INTENT_MAIN_BACKUP && resultCode == RESULT_OK) {
             if (intent.getBooleanExtra("reload", false)) {
@@ -397,6 +464,10 @@ public class MainActivity extends BaseActivity
                     authKey = intent.getByteArrayExtra(Constants.EXTRA_AUTH_PASSWORD_KEY);
 
                 updateEncryption(authKey);
+            }
+        } else if (requestCode == Constants.INTENT_MAIN_QR_OPEN_IMAGE && resultCode == RESULT_OK) {
+            if (intent != null) {
+                addQRCode(ScanQRCodeFromFile.scanQRImage(this, intent.getData()));
             }
         }
     }
@@ -435,12 +506,18 @@ public class MainActivity extends BaseActivity
             if (mode == SortMode.UNSORTED) {
                 sortMenu.setIcon(R.drawable.ic_sort_inverted_white);
                 menu.findItem(R.id.menu_sort_none).setChecked(true);
+            } else if (mode == SortMode.ISSUER) {
+                sortMenu.setIcon(R.drawable.ic_sort_inverted_label_white);
+                menu.findItem(R.id.menu_sort_issuer).setChecked(true);
             } else if (mode == SortMode.LABEL) {
                 sortMenu.setIcon(R.drawable.ic_sort_inverted_label_white);
                 menu.findItem(R.id.menu_sort_label).setChecked(true);
             } else if (mode == SortMode.LAST_USED) {
                 sortMenu.setIcon(R.drawable.ic_sort_inverted_time_white);
                 menu.findItem(R.id.menu_sort_last_used).setChecked(true);
+            } else if (mode == SortMode.MOST_USED) {
+                sortMenu.setIcon(R.drawable.ic_sort_inverted_time_white);
+                menu.findItem(R.id.menu_sort_most_used).setChecked(true);
             }
         }
 
@@ -454,11 +531,7 @@ public class MainActivity extends BaseActivity
 
             @Override
             public boolean onQueryTextChange(String newText) {
-                if (newText.isEmpty())
-                    adapter.filterByTags(tagsDrawerAdapter.getActiveTags());
-                else
-                    adapter.getFilter().filter(newText);
-
+                setFilterString(newText);
                 return false;
             }
         });
@@ -490,6 +563,15 @@ public class MainActivity extends BaseActivity
         return true;
     }
 
+    private void setFilterString(String newText) {
+        if (newText.isEmpty())
+            adapter.filterByTags(tagsDrawerAdapter.getActiveTags());
+        else
+            adapter.getFilter().filter(newText);
+
+        this.filterString = newText;
+    }
+
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         int id = item.getItemId();
@@ -515,6 +597,14 @@ public class MainActivity extends BaseActivity
                 adapter.setSortMode(SortMode.UNSORTED);
                 touchHelperCallback.setDragEnabled(true);
             }
+        } else if (id == R.id.menu_sort_issuer) {
+            item.setChecked(true);
+            sortMenu.setIcon(R.drawable.ic_sort_inverted_label_white);
+            saveSortMode(SortMode.ISSUER);
+            if(adapter != null) {
+                adapter.setSortMode(SortMode.ISSUER);
+                touchHelperCallback.setDragEnabled(false);
+            }
         } else if (id == R.id.menu_sort_label) {
             item.setChecked(true);
             sortMenu.setIcon(R.drawable.ic_sort_inverted_label_white);
@@ -531,8 +621,18 @@ public class MainActivity extends BaseActivity
                 adapter.setSortMode(SortMode.LAST_USED);
                 touchHelperCallback.setDragEnabled(false);
             }
-            if (! settings.getLastUsedDialogShown())
-                showLastUsedDialog();
+            if (! settings.getUsedTokensDialogShown())
+                showUsedTokensDialog();
+        } else if (id == R.id.menu_sort_most_used) {
+            item.setChecked(true);
+            sortMenu.setIcon(R.drawable.ic_sort_inverted_time_white);
+            saveSortMode(SortMode.MOST_USED);
+            if (adapter != null) {
+                adapter.setSortMode(SortMode.MOST_USED);
+                touchHelperCallback.setDragEnabled(false);
+            }
+            if (! settings.getUsedTokensDialogShown())
+                showUsedTokensDialog();
         } else if (tagsToggle.onOptionsItemSelected(item)) {
             return true;
         }
@@ -540,17 +640,11 @@ public class MainActivity extends BaseActivity
         return super.onOptionsItemSelected(item);
     }
 
-    private void showLastUsedDialog() {
+    private void showUsedTokensDialog() {
         AlertDialog.Builder builder = new AlertDialog.Builder(this);
-        builder.setTitle(R.string.dialog_title_manual_entry)
-                .setTitle(R.string.dialog_title_last_used)
-                .setMessage(R.string.dialog_msg_last_used)
-                .setPositiveButton(android.R.string.ok, new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialogInterface, int i) {
-                        settings.setLastUsedDialogShown(true);
-                    }
-                })
+        builder.setTitle(R.string.dialog_title_used_tokens)
+                .setMessage(R.string.dialog_msg_used_tokens)
+                .setPositiveButton(android.R.string.ok, (DialogInterface dialogInterface, int i) -> settings.setUsedTokensDialogShown(true))
                 .create()
                 .show();
     }
@@ -591,10 +685,13 @@ public class MainActivity extends BaseActivity
                 settings.setAllTagsToggle(checkedTextView.isChecked());
 
                 for(int i = 0; i < tagsDrawerListView.getChildCount(); i++) {
-                    CheckedTextView childCheckBox = (CheckedTextView)tagsDrawerListView.getChildAt(i);
+                    CheckedTextView childCheckBox = (CheckedTextView) tagsDrawerListView.getChildAt(i);
                     childCheckBox.setChecked(checkedTextView.isChecked());
-                    tagsDrawerAdapter.setTagState(childCheckBox.getText().toString(), childCheckBox.isChecked());
-                    settings.setTagToggle(childCheckBox.getText().toString(), childCheckBox.isChecked());
+                }
+
+                for (String tag: tagsDrawerAdapter.getTags()) {
+                    tagsDrawerAdapter.setTagState(tag, checkedTextView.isChecked());
+                    settings.setTagToggle(tag, checkedTextView.isChecked());
                 }
 
                 if(checkedTextView.isChecked()) {
@@ -684,5 +781,88 @@ public class MainActivity extends BaseActivity
         }
         tagsDrawerAdapter.setTags(tagsHashMap);
         adapter.filterByTags(tagsDrawerAdapter.getActiveTags());
+    }
+
+    @Override
+    public void onUserInteraction(){
+        super.onUserInteraction();
+
+        // Refresh Blackout Timer
+        if (countDownTimer != null)
+            countDownTimer.cancel();
+
+        if (setCountDownTimerNow())
+            countDownTimer.start();
+    }
+
+    private boolean setCountDownTimerNow() {
+        int secondsToBlackout = 1000 * settings.getAuthInactivityDelay();
+
+        if (settings.getAuthMethod() == AuthMethod.NONE || !settings.getAuthInactivity() || secondsToBlackout == 0)
+            return false;
+
+        countDownTimer = new CountDownTimer(secondsToBlackout, 1000) {
+            @Override
+            public void onTick(long millisUntilFinished) {
+            }
+
+            @Override
+            public void onFinish() {
+                authenticate(R.string.auth_msg_authenticate);
+                this.cancel();
+            }
+        };
+
+        return true;
+    }
+
+    private void openFileWithPermissions(int intentId){
+        if (ContextCompat.checkSelfPermission(this, Manifest.permission.READ_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
+            showOpenFileSelector(intentId);
+        } else {
+            ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.READ_EXTERNAL_STORAGE}, Constants.PERMISSIONS_MAIN_QR_READ_IMAGE);
+        }
+    }
+
+    private void showOpenFileSelector(int intentId){
+        Intent intent = new Intent(Intent.ACTION_OPEN_DOCUMENT);
+        intent.addCategory(Intent.CATEGORY_OPENABLE);
+        intent.setType("image/*");
+        startActivityForResult(intent, intentId);
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        if (requestCode == Constants.PERMISSIONS_MAIN_QR_READ_IMAGE) {
+            if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                showOpenFileSelector(Constants.INTENT_MAIN_QR_OPEN_IMAGE);
+            } else {
+                Toast.makeText(this, R.string.backup_toast_storage_permissions, Toast.LENGTH_LONG).show();
+            }
+        } else {
+            super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        }
+    }
+    
+    private void addQRCode(String result){
+        if(!TextUtils.isEmpty(result)) {
+            try {
+                Entry e = new Entry(result);
+                e.updateOTP();
+                e.setLastUsed(System.currentTimeMillis());
+                adapter.addEntry(e);
+                refreshTags();
+            } catch (Exception e) {
+                Toast.makeText(this, R.string.toast_invalid_qr_code, Toast.LENGTH_LONG).show();
+            }
+        }
+    }
+
+    private class ProcessLifecycleObserver implements DefaultLifecycleObserver {
+        @Override
+        public void onStop(LifecycleOwner owner) {
+            if (MainActivity.this.settings.getRelockOnBackground())
+                MainActivity.this.requireAuthentication = true;
+        }
     }
 }
