@@ -24,11 +24,13 @@ package org.shadowice.flocke.andotp.Activities;
 
 import android.content.res.Configuration;
 import android.content.res.Resources;
+import android.os.Build;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import org.shadowice.flocke.andotp.R;
 import org.shadowice.flocke.andotp.Utilities.Settings;
+import org.shadowice.flocke.andotp.Utilities.Tools;
 
 import java.util.Locale;
 
@@ -39,8 +41,13 @@ public abstract class ThemedActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         settings = new Settings(this);
 
-        setTheme();
+        setTheme(settings.getTheme());
         setLocale();
+
+        //Set navigation bar color
+        if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+            getWindow().setNavigationBarColor(Tools.getThemeColor(this,R.attr.navigationBarColor));
+        }
 
         super.onCreate(savedInstanceState);
     }
@@ -50,18 +57,6 @@ public abstract class ThemedActivity extends AppCompatActivity {
         setLocale();
 
         super.onResume();
-    }
-
-    public void setTheme() {
-        String theme = settings.getTheme();
-
-        if (theme.equals("light")) {
-            setTheme(R.style.AppTheme_NoActionBar);
-        } else if (theme.equals("dark")) {
-            setTheme(R.style.AppTheme_Dark_NoActionBar);
-        } else if (theme.equals("black")) {
-            setTheme(R.style.AppTheme_Black_NoActionBar);
-        }
     }
 
     public void setLocale() {
