@@ -28,7 +28,7 @@ import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.graphics.Canvas;
 import android.graphics.drawable.Drawable;
-import android.support.v7.app.AppCompatDelegate;
+import androidx.appcompat.app.AppCompatDelegate;
 
 import org.shadowice.flocke.andotp.R;
 
@@ -40,24 +40,29 @@ public class EntryThumbnail {
 
     public enum EntryThumbnails {
         Default(R.mipmap.ic_launcher_round),
+        OneAndOne(R.drawable.thumb_1and1),
         OnePassword(R.drawable.thumb_1password),
         Adobe(R.drawable.thumb_adobe),
+        AdGuard(R.drawable.thumb_adguard),
         Airbrake(R.drawable.thumb_airbrake),
         AllegroPl(R.drawable.thumb_allegropl),
         Amazon(R.drawable.thumb_amazon),
         AmazonWebServices(R.drawable.thumb_amazonwebservices),
         AngelList(R.drawable.thumb_angellist),
         AnimeBytes(R.drawable.thumb_animebytes),
+        Apache(R.drawable.thumb_apache),
         Apple(R.drawable.thumb_apple),
         ArenaNet(R.drawable.thumb_arenanet),
         Atlassian(R.drawable.thumb_atlassian),
         AVM(R.drawable.thumb_avm),
         Backblaze(R.drawable.thumb_backblaze),
         BattleNet(R.drawable.thumb_battlenet),
+        Betterment(R.drawable.thumb_betterment),
         Binance(R.drawable.thumb_binance),
         BitBucket(R.drawable.thumb_bitbucket),
         Bitcoin(R.drawable.thumb_bitcoin),
         Bitfinex(R.drawable.thumb_bitfinex),
+        Bitmex(R.drawable.thumb_bitmex),
         Bitstamp(R.mipmap.thumb_bitstamp, AssetType.Bitmap),
         Bitpanda(R.drawable.thumb_bitpanda),
         Bittrex(R.drawable.thumb_bittrex),
@@ -67,17 +72,24 @@ public class EntryThumbnail {
         BlockchainInfo(R.drawable.thumb_blockchain_info),
         CloudDownload(R.drawable.thumb_cloud_download),
         Cloudflare(R.drawable.thumb_cloudflare),
+        Cobranded(R.drawable.thumb_cobranded),
         Codegiant(R.drawable.thumb_codegiant),
         Coinbase(R.drawable.thumb_coinbase),
         ComputerBase(R.drawable.thumb_computerbase),
         ConnectWiseManage(R.drawable.thumb_connectwise_manage),
         CozyCloud(R.drawable.thumb_cozycloud),
+        Crowdin(R.drawable.thumb_crowdin),
+        Dashlane(R.drawable.thumb_dashlane),
         Debian(R.drawable.thumb_debian),
         Degiro(R.drawable.thumb_degiro),
+        Diaspora(R.drawable.thumb_diaspora),
+        Digidentity(R.drawable.thumb_digidentity),
         DigitalOcean(R.drawable.thumb_digital_ocean),
         Discord(R.drawable.thumb_discord),
         Discourse(R.drawable.thumb_discourse),
         Disroot(R.drawable.thumb_disroot),
+        Docker(R.drawable.thumb_docker),
+        DocuSign(R.drawable.thumb_docusign),
         Dropbox(R.drawable.thumb_dropbox),
         ElectronicArts(R.drawable.thumb_electronic_arts),
         Email(R.drawable.thumb_email),
@@ -86,9 +98,11 @@ public class EntryThumbnail {
         Facebook(R.drawable.thumb_facebook),
         FACEIT(R.drawable.thumb_faceit),
         Fingerprint(R.drawable.thumb_fingerprint),
+        Finnair(R.drawable.thumb_finnair),
         Firefox(R.drawable.thumb_firefox),
         Flight(R.drawable.thumb_flight_takeoff),
         Floatplane(R.drawable.thumb_floatplane),
+        Fritz(R.drawable.thumb_fritz),
         Gamepad(R.drawable.thumb_gamepad),
         Gandi(R.drawable.thumb_gandi),
         Git(R.drawable.thumb_git),
@@ -106,22 +120,28 @@ public class EntryThumbnail {
         HomeAssistant(R.drawable.thumb_home_assistant),
         HumbleBundle(R.drawable.thumb_humblebundle),
         HurricaneElectric(R.drawable.thumb_hurricane_electric),
+        IBM(R.drawable.thumb_ibm),
         Iconomi(R.drawable.thumb_iconomi),
         IFTTT(R.drawable.thumb_ifttt),
         Instagram(R.drawable.thumb_instagram),
         INWX(R.drawable.thumb_inwx),
         Itchio(R.drawable.thumb_itchio),
         Jagex(R.drawable.thumb_jagex),
+        Joomla(R.drawable.thumb_joomla),
         Kickstarter(R.drawable.thumb_kickstarter),
         Kraken(R.drawable.thumb_kraken),
         Kucoin(R.drawable.thumb_kucoin),
         LastPass(R.drawable.thumb_lastpass),
+        LibreNMS(R.drawable.thumb_librenms),
+        Lichess(R.drawable.thumb_lichess),
+        LinkedIn(R.drawable.thumb_linkedin),
         Linode(R.drawable.thumb_linode),
         Liqui(R.drawable.thumb_liqui),
         LocalBitcoins(R.drawable.thumb_localbitcoins),
         LocalMonero(R.drawable.thumb_localmonero),
         LogMeIn(R.drawable.thumb_logmein),
         Mailbox(R.drawable.thumb_mailbox),
+        Mailcow(R.drawable.thumb_mailcow),
         Mailgun(R.drawable.thumb_mailgun),
         Mapbox(R.drawable.thumb_mapbox),
         Mastodon(R.drawable.thumb_mastodon),
@@ -133,6 +153,8 @@ public class EntryThumbnail {
         Migadu(R.drawable.thumb_migadu),
         Miraheze(R.drawable.thumb_miraheze),
         Mixer(R.drawable.thumb_mixer),
+        NameCheap(R.drawable.thumb_namecheap),
+        NameCom(R.drawable.thumb_namecom),
         NAS(R.drawable.thumb_nas),
         netcup(R.drawable.thumb_netcup),
         NextCloud(R.drawable.thumb_nextcloud),
@@ -142,18 +164,27 @@ public class EntryThumbnail {
         Okta(R.drawable.thumb_okta),
         OnlineNet(R.drawable.thumb_online),
         OpenVZ(R.drawable.thumb_openvz),
+        OPNsense(R.drawable.thumb_opnsense),
         Origin(R.drawable.thumb_origin),
         OVH(R.drawable.thumb_ovh),
         Packet(R.drawable.thumb_packet),
         Patreon(R.drawable.thumb_patreon),
         PayPal(R.drawable.thumb_paypal),
+        PaySafe(R.drawable.thumb_paysafecard),
+        PayWithPrivacy(R.drawable.thumb_paywithprivacy),
         PCloud(R.drawable.thumb_pcloud),
         Phabricator(R.drawable.thumb_phabricator),
         phpMyAdmin(R.drawable.thumb_phpmyadmin),
         Plurk(R.drawable.thumb_plurk),
+        Posteo(R.drawable.thumb_posteo),
+        PrivateInternetAccess(R.drawable.thumb_private_internet_access),
         ProtonMail(R.drawable.thumb_protonmail),
+        Proxmox(R.drawable.thumb_proxmox),
+        PyPI(R.drawable.thumb_pypi),
         Rackspace(R.drawable.thumb_rackspace),
         Reddit(R.drawable.thumb_reddit),
+        RipeNNC(R.drawable.thumb_ripe_ncc),
+        Robinhood(R.drawable.thumb_robinhood),
         Rockstar(R.drawable.thumb_rockstar),
         RSS(R.drawable.thumb_rss),
         SAP(R.drawable.thumb_sap),
@@ -176,11 +207,13 @@ public class EntryThumbnail {
         TeaHub(R.drawable.thumb_teahub),
         TeamViewer(R.drawable.thumb_teamviewer),
         Terminal(R.drawable.thumb_terminal),
+        TransIP(R.drawable.thumb_transip),
         Trello(R.drawable.thumb_trello),
         Tumblr(R.drawable.thumb_tumblr),
         Tutanota(R.drawable.thumb_tutanota),
         Twitch(R.drawable.thumb_twitch),
         Twitter(R.drawable.thumb_twitter),
+        Uber(R.drawable.thumb_uber),
         Ubisoft(R.drawable.thumb_ubisoft),
         UbiquitiNetworks(R.drawable.thumb_ubnt),
         UbuntuOne(R.drawable.thumb_ubuntu_one),
@@ -190,6 +223,7 @@ public class EntryThumbnail {
         Wallabag(R.drawable.thumb_wallabag),
         Wallet(R.drawable.thumb_wallet),
         Wargaming(R.drawable.thumb_wargaming),
+        Wasabi(R.drawable.thumb_wasabi),
         Wikimedia(R.drawable.thumb_wikimedia),
         Wordpress(R.drawable.thumb_wordpress),
         Zoho(R.drawable.thumb_zoho);
@@ -213,14 +247,21 @@ public class EntryThumbnail {
         public AssetType getAssetType() {
             return assetType;
         }
+
+        public static EntryThumbnails valueOfIgnoreCase(String thumbnail) {
+            for (EntryThumbnails entryThumbnails : values())
+                if (entryThumbnails.name().equalsIgnoreCase(thumbnail)) return entryThumbnails;
+                throw new IllegalArgumentException();
+        }
     }
 
-    public static Bitmap getThumbnailGraphic(Context context, String label, int size, EntryThumbnails thumbnail) {
+    public static Bitmap getThumbnailGraphic(Context context, String issuer, String label, int size, EntryThumbnails thumbnail) {
         AppCompatDelegate.setCompatVectorFromResourcesEnabled(true);
 
         if (thumbnail == EntryThumbnails.Default && size > 0) {
             LetterBitmap letterBitmap = new LetterBitmap(context);
-            return letterBitmap.getLetterTile(label, label, size, size);
+            String letterSrc = issuer.isEmpty() ? label : issuer;
+            return letterBitmap.getLetterTile(letterSrc, letterSrc, size, size);
         } else if (thumbnail != EntryThumbnails.Default) {
 
             try {
